% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mood_distribution.R
\name{plot_mood_distribution}
\alias{plot_mood_distribution}
\title{Plot Mood Distribution}
\usage{
plot_mood_distribution(moods)
}
\arguments{
\item{moods}{A character vector containing "positive", "neutral", or "negative".}
}
\value{
A ggplot2 object.
}
\description{
Creates a bar chart of the count of detected sentiments.
}
\examples{
moods <- detect_mood(c("I love it", "I hate this", "meh"))
plot_mood_distribution(moods)
}
